coding #3 moviendo clases a nueva estructura de proyecto

+++ services/api/src/main/groovy/com/makingdevs/Concepto.groovy
@@ -1,7 +1,7 @@
 package com.makingdevs
 
 class Concepto{
-  Float cantidad
+  BigDecimal cantidad
   String unidad
   String noIdentificacion
   String descripcion
